Fix notification truncation for ActiveSupport 3.x.x. ActiveSupport 3.x.x will escape UTF8 characters and some HTML entities by default, which meant that the code to truncate notification alerts wouldn't work as designed.
E.g., with AS 3.x.x "é" would be encoded to \u00e9 which is six
bytes, not two. This was changed for AS 4.x.x here:
https://github.com/rails/rails/commit/80e7552073712e102c584cfc54cb3eff2c1f0f52.

Separate tests are provided for the different versions of
ActiveSupport. Tests can be run for AS 3.x.x by temporarily changing
the gemspec version requirement, running `bundle`, then running the
specs.

This commit also refactors the truncation and encoding related logic
into a new class.